# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get project summary stats
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /projects/{projectId}/statuses:
    get:
      operationId: listProjectStatuses
      tags: [statuses]
      summary: List project statuses
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of project statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectStatus"
    put:
      operationId: updateProjectStatuses
      tags: [statuses]
      summary: Replace project statuses
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectStatusesBody"
      responses:
        "200":
          description: Updated project statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectStatus"
...
  /flow-sessions/start:
    post:
      operationId: startFlowSession
      tags: [flow-sessions]
      summary: Start a flow session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartFlowSessionRequest"
      responses:
        "200":
          description: Started flow session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartFlowSessionResponse"
  /flow-sessions/{id}/stop:
    post:
      operationId: stopFlowSession
      tags: [flow-sessions]
      summary: Stop a flow session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopFlowSessionRequest"
      responses:
        "200":
          description: Stopped flow session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopFlowSessionResponse"
  /daily-review:
    post:
      operationId: saveDailyReview
      tags: [daily-review]
      summary: Save daily review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveDailyReviewRequest"
      responses:
        "200":
          description: Saved daily review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveDailyReviewResponse"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
          description: Short uppercase identifier e.g. "PROJ"
        description:
          type: string
        color:
          type: string
        issueCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, key, color, issueCount, createdAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        key:
          type: string
        description:
          type: string
        color:
          type: string
        statuses:
          type: array
          items:
            $ref: "#/components/schemas/ProjectStatusInput"
      required: [name, key]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: string
        totalIssues:
          type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
        byPriority:
          type: object
          additionalProperties:
            type: integer
        byType:
          type: object
          additionalProperties:
            type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
      required: [projectId, totalIssues, byStatus, byPriority, byType, recentActivity]
    ProjectStatus:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        color:
          type: string
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, name, color, position, createdAt]
    ProjectStatusInput:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        position:
          type: integer
      required: [name, color, position]
    UpdateProjectStatusesBody:
      type: object
      properties:
        statuses:
          type: array
          items:
            $ref: "#/components/schemas/ProjectStatusInput"
      required: [statuses]
    Issue:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        issueKey:
          type: string
          description: e.g. "PROJ-42"
        title:
          type: string
        status:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [task, bug, feature, story]
        assignee:
          type: string
          nullable: true
        reporter:
          type: string
        labels:
          type: array
          items:
            type: string
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, issueKey, title, status, priority, type, reporter, labels, commentCount, createdAt, updatedAt]
...
    FlowSession:
      type: object
      properties:
        id:
          type: string
        issueId:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, startedAt, createdAt, updatedAt]
    DailyReview:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        summary:
          type: string
          nullable: true
        completedIssueIds:
          type: array
          items:
            type: string
        carriedIssueIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, date, completedIssueIds, carriedIssueIds, createdAt, updatedAt]
    StartFlowSessionRequest:
      type: object
      properties:
        issueId:
          type: string
      required: [issueId]
    StartFlowSessionResponse:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/FlowSession"
      required: [session]
    StopFlowSessionRequest:
      type: object
      properties:
        note:
          type: string
    StopFlowSessionResponse:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/FlowSession"
      required: [session]
    SaveDailyReviewRequest:
      type: object
      properties:
        date:
          type: string
        summary:
          type: string
        completedIssueIds:
          type: array
          items:
            type: string
        carriedIssueIds:
          type: array
          items:
            type: string
      required: [date]
    SaveDailyReviewResponse:
      type: object
      properties:
        review:
          $ref: "#/components/schemas/DailyReview"
      required: [review]